+++ src/index.js
@@ -18,6 +18,7 @@ export { default as Title } from './title';
 export { default as DateEasily } from './date-easily';
 export { default as EditTable } from './edit-table';
 export { default as Ellipsis } from './ellipsis';
+export { default as InnerHtml } from './inner-html';
 
 // 高耦合组件
 export { default as localConfig } from './local-config';