+++ src/inner-html/index.js
@@ -1,0 +1,8 @@
+import React from 'react';
+
+export default function InnerHtml({
+  html = '',
+  ...rest
+}) {
+  return <div dangerouslySetInnerHTML={{ __html: html }} {...rest} />;
+}